% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateExtendedEdgeListfromMatrices.R
\name{GenerateExtendedEdgeListfromMatrices}
\alias{GenerateExtendedEdgeListfromMatrices}
\title{Generate Extended Edges List that is Compatible with MuxViz}
\usage{
GenerateExtendedEdgeListfromMatrices(
  Matrix_list,
  Layers,
  Nodes,
  Threshold_percentile = 0,
  doOMST = FALSE,
  ...
)
}
\arguments{
\item{Matrix_list}{List of matrices to be included in the multilayer network}

\item{Layers}{Number of Layers in Network}

\item{Nodes}{Number of Nodes in Network}

\item{Threshold_percentile}{Percentile at which to threshold each matrix}

\item{doOMST}{Perform OMST thresholding}

\item{...}{If stucturally constrained, Bin_struct = Bin_struct (case sensitive)}
}
\value{
Returns Extended Edges List to be used with muxViz
}
\description{
Generate Extended Edges List that is Compatible with MuxViz
}
\examples{
Ext_edges <- GenerateExtendedEdgeListfromMatrices <- function(Matrix_list, Layers, Nodes, Threshold_percentile = 0,  doOMST = FALSE, Bin_struct = Bin_struct)


}
